Better dataloader configuration
