+++ examples/nanoChatGPT/config/train.yaml
@@ -6,7 +6,7 @@ eval_iters: 20
 always_save_checkpoint: False  # if True, always save a checkpoint after each eval
 init_from: gpt2  # scratch or resume or gpt2*
 # data
-dataset: shakespeare
+dataset: tldr
 gradient_accumulation_steps: 10  # used to simulate larger batch sizes
 batch_size: 12  # if gradient_accumulation_steps > 1, this is the micro-batch size
 block_size: 1024
