+++ examples/nanoChatGPT/config/train_rlhf.yaml
@@ -10,7 +10,7 @@ always_save_checkpoint: True # if True, always save a checkpoint after each eval
 init_from: resume # 'scratch' or 'resume' or 'gpt2*'
 init_reward_from: resume
 # data:
-dataset: shakespeare # 'openwebtext', 'shakespeare', 'openai_summarize_tldr'
+dataset: tldr # 'openwebtext', 'shakespeare', 'tldr'
 gradient_accumulation_steps: 10 # used to simulate larger batch sizes
 batch_size: 40 # if gradient_accumulation_steps > 1, this is the micro-batch size
 block_size: 1024
