--- examples/nanoChatGPT/data/shakespeare.py
@@ -1,24 +1,0 @@
-from pathlib import Path
-
-from .utils import Collate, create_infinite_dataloader, PairedDataset
-
-HERE = Path(__file__).parent
-
-
-def create_datasets(config):
-    data_dir = HERE.parent / "models" / "nanoGPT" / "data" / config["dataset"]
-    train_data = PairedDataset(data_dir / "train.bin", block_size=config["block_size"])
-    val_data = PairedDataset(data_dir / "val.bin", block_size=config["block_size"])
-
-    return train_data, val_data
-
-
-def get_dataloaders(config):
-    train_data, val_data = create_datasets(config)
-
-    train_loader = create_infinite_dataloader(
-        train_data, config, Collate(config["device"])
-    )
-    val_loader = create_infinite_dataloader(val_data, config, Collate(config["device"]))
-
-    return train_loader, val_loader
