+++ examples/nanoChatGPT/data/utils.py
@@ -1,10 +1,4 @@
-from typing import Optional
-
-import numpy as np
-import torch
-import torch.nn as nn
-from tensordict import tensorclass
-from torch.utils.data import DataLoader, Dataset
+from torch.utils.data import DataLoader
 
 
 def create_infinite_dataloader(data, config, collate_fn):
@@ -15,54 +9,9 @@ def create_infinite_dataloader(data, config, collate_fn):
     dl = DataLoader(
         data,
         batch_size=config["batch_size"],
-        shuffle=True,  # TODO: perhaps validation set shouldn't be shuffled?
+        shuffle=True,
         collate_fn=collate_fn,
         drop_last=True,
     )
     while True:
         yield from dl
-
-
-class Collate(nn.Module):
-    def __init__(self, device="cpu"):
-        super().__init__()
-        self.device = torch.device(device)
-
-    def __call__(self, batch):
-        batch = torch.stack(batch, dim=0).contiguous()
-        batch.batch_size = []
-        if self.device.type == "cuda":
-            batch = batch.pin_memory()
-        return batch.to(self.device)
-
-
-@tensorclass
-class Data:
-    prompt: torch.Tensor
-    target: torch.Tensor
-    logits: Optional[torch.Tensor] = None
-    loss: Optional[torch.Tensor] = None
-
-
-class PairedDataset(Dataset):
-    def __init__(self, path, block_size):
-        self._memmap = np.memmap(path, dtype=np.uint16, mode="r")
-        self.block_size = block_size
-
-    def __getitem__(self, idx):
-        return Data(
-            prompt=torch.from_numpy(
-                self._memmap[idx : idx + self.block_size].astype(np.int64)
-            ),
-            target=torch.from_numpy(
-                self._memmap[idx + 1 : idx + self.block_size + 1].astype(np.int64)
-            ),
-            batch_size=[self.block_size],
-        )
-
-    def __len__(self):
-        # how many sequences of length block_size + 1 can we extract from the data?
-        # the valid starting points for such a sequence are those tokens that aren't in
-        # the final block_size positions. so it's just the length of the overall
-        # sequence minus the block_size
-        return len(self._memmap) - self.block_size
