+++ examples/nanoChatGPT/train_reward.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import torch
 
-from data.openai_summarize_comparisons import get_dataloaders
+from data import get_reward_dataloaders
 from models.reward import init_reward_model
 from shared import create_lr_scheduler, setup
 from utils import load_and_update_config
@@ -43,7 +43,7 @@ def main():
 
     estimate_loss = create_loss_estimator(config, ctx)
 
-    train_loader, val_loader = get_dataloaders(config)
+    train_loader, val_loader = get_reward_dataloaders(config)
 
     # these will already have been set if resuming from previous checkpoint
     iter_num = config.setdefault("iter_num", 0)
